+++ core/src/main/java/org/sterl/spring/persistent_tasks/scheduler/SchedulerService.java
@@ -2,13 +2,13 @@ package org.sterl.spring.persistent_tasks.scheduler;
 
 import java.io.Serializable;
 import java.time.OffsetDateTime;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.Future;
 
+import org.slf4j.event.Level;
 import org.springframework.lang.NonNull;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
@@ -18,6 +18,7 @@ import org.springframework.transaction.support.TransactionTemplate;
 import org.sterl.spring.persistent_tasks.api.AddTriggerRequest;
 import org.sterl.spring.persistent_tasks.api.TriggerKey;
 import org.sterl.spring.persistent_tasks.scheduler.component.EditSchedulerStatusComponent;
+import org.sterl.spring.persistent_tasks.scheduler.component.RunOrQueueComponent;
 import org.sterl.spring.persistent_tasks.scheduler.component.TaskExecutorComponent;
 import org.sterl.spring.persistent_tasks.scheduler.entity.SchedulerEntity;
 import org.sterl.spring.persistent_tasks.trigger.TriggerService;
@@ -43,16 +44,18 @@ public class SchedulerService {
     @Getter
     private final String name;
     private final TriggerService triggerService;
+    
     private final TaskExecutorComponent taskExecutor;
     private final EditSchedulerStatusComponent editSchedulerStatus;
+
+    private final RunOrQueueComponent runOrQueue;
+
     private final TransactionTemplate trx;
-    private final Map<Long, TriggerEntity> shouldRun = new ConcurrentHashMap<>();
 
     @PostConstruct
     public void start() {
         taskExecutor.start();
-        final var s = editSchedulerStatus.checkinToRegistry(name);
-        log.info("Started {} on {} threads.", s, taskExecutor.getMaxThreads());
+        editSchedulerStatus.checkinToRegistry(name, 0, taskExecutor.getMaxThreads());
     }
 
     public void setMaxThreads(int value) {
@@ -63,7 +66,6 @@ public class SchedulerService {
     public void stop() {
         taskExecutor.close();
         editSchedulerStatus.offline(name);
-        log.info("Stopped {}", name);
     }
 
     public void shutdownNow() {
@@ -73,22 +75,15 @@ public class SchedulerService {
         editSchedulerStatus.offline(name);
     }
 
-    @Transactional
-    public SchedulerEntity pingRegistry() {
-        var result = editSchedulerStatus.checkinToRegistry(name);
-        result.setRunnungTasks(taskExecutor.getRunningTasks());
-        result.setTasksSlotCount(taskExecutor.getMaxThreads());
-        log.debug("Ping {}", result);
-        return result;
-    }
-
     public SchedulerEntity getScheduler() {
         var result = editSchedulerStatus.get(name);
         return result;
     }
 
     public Optional<SchedulerEntity> findStatus(String name) {
-        return editSchedulerStatus.find(name);
+        if (name == null) return Optional.empty();
+        else if (name.equals(this.name)) return Optional.of(getScheduler());
+        else return editSchedulerStatus.find(name);
     }
 
     /**
@@ -112,17 +107,18 @@ public class SchedulerService {
         if (taskExecutor.getFreeThreads() > 0) {
             final var result = trx.execute(t -> {
                 var triggers = triggerService.lockNextTrigger(name, taskExecutor.getFreeThreads(), timeDue);
-                pingRegistry().addRunning(triggers.size());
+                editSchedulerStatus.checkinToRegistry(name, 
+                        taskExecutor.countRunning() + triggers.size(), taskExecutor.getMaxThreads());
                 return triggers;
             });
 
             return taskExecutor.submit(result);
         } else {
-            log.debug("No free threads {}/{} right now to run jobs due for: {}",
+            log.info("No free threads {}/{} right now to run jobs due for: {}",
                     taskExecutor.getFreeThreads(),
                     taskExecutor.getMaxThreads(),
                     timeDue);
-            pingRegistry();
+            editSchedulerStatus.checkinToRegistry(name, taskExecutor.countRunning(), taskExecutor.getMaxThreads());
             return Collections.emptyList();
         }
     }
@@ -136,31 +132,14 @@ public class SchedulerService {
      */
     @Transactional(timeout = 10)
     public <T extends Serializable> TriggerKey runOrQueue(AddTriggerRequest<T> triggerRequest) {
-        var trigger = triggerService.queue(triggerRequest);
-
-        if (!trigger.shouldRunInFuture()) {
-            if (taskExecutor.getFreeThreads() > 0) {
-                trigger = triggerService.markTriggersAsRunning(trigger, name);
-                pingRegistry().addRunning(1);
-                shouldRun.put(trigger.getId(), trigger);
-                log.debug("{} added for immediate execution, waitng for commit on={}", trigger.getKey(), name);
-            } else {
-                log.debug("Currently not enough free thread available {} of {} in use. PersistentTask {} queued.",
-                        taskExecutor.getFreeThreads(), taskExecutor.getMaxThreads(), trigger.getKey());
-            }
-        }
-        // we will listen for the commit event to execute this trigger ...
-        return trigger.getKey();
+        return runOrQueue.execute(triggerRequest);
     }
 
     @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
-    void checkIfTrigerIsRunning(TriggerAddedEvent addedTrigger) {
-        final var toRun = shouldRun.remove(addedTrigger.id());
-        if (toRun != null) {
-            taskExecutor.submit(toRun);
-            log.debug("{} immediately started on={}.", addedTrigger.key(), name);
+    public void checkIfTrigerShouldRun(TriggerAddedEvent addedTrigger) {
+        if (runOrQueue.checkIfTrigerShouldRun(addedTrigger.id())) {
+            editSchedulerStatus.checkinToRegistry(name, taskExecutor.countRunning(), taskExecutor.getMaxThreads());
         }
-        // TODO implement a cleanup for old pending triggers which may never been triggered!
     }
 
     public SchedulerEntity getStatus() {
@@ -175,7 +154,7 @@ public class SchedulerService {
                 .toList();
 
         int running = triggerService.markTriggersAsRunning(runningKeys, name);
-        log.debug("({}) - {} trigger(s) are running on {} schedulers", 
+        log.atLevel(running > 0 ? Level.INFO : Level.DEBUG).log("({}) - {} trigger(s) are running on {} schedulers", 
                 running, runningKeys, schedulers);
         return triggerService.rescheduleAbandonedTasks(timeout);
     }
@@ -183,8 +162,11 @@ public class SchedulerService {
     public List<SchedulerEntity> listAll() {
         return editSchedulerStatus.listAll();
     }
-    
+    public Collection<Future<TriggerKey>> getRunning() {
+        return taskExecutor.getRunningTasks();
+    }
     public boolean hasRunningTriggers() {
-        return !this.taskExecutor.getRunningTriggers().isEmpty();
+        return !taskExecutor.isStopped() 
+                && (taskExecutor.countRunning() > 0 || runOrQueue.hasWaitingTriggers());
     }
 }