+++ core/src/main/java/org/sterl/spring/persistent_tasks/scheduler/component/TaskExecutorComponent.java
@@ -3,6 +3,7 @@ package org.sterl.spring.persistent_tasks.scheduler.component;
 import java.io.Closeable;
 import java.time.Duration;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
@@ -20,15 +21,20 @@ import org.sterl.spring.persistent_tasks.api.TriggerKey;
 import org.sterl.spring.persistent_tasks.trigger.TriggerService;
 import org.sterl.spring.persistent_tasks.trigger.model.TriggerEntity;
 
-import jakarta.annotation.PostConstruct;
-import jakarta.annotation.PreDestroy;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
 
+/**
+ * The executor of a scheduler
+ * <p>
+ * Not a spring bean!
+ * </p>
+ */
 @Slf4j
 public class TaskExecutorComponent implements Closeable {
 
+    private final String schedulerName;
     private final TriggerService triggerService;
     private final AtomicInteger maxThreads = new AtomicInteger(0);
     @Getter
@@ -36,18 +42,21 @@ public class TaskExecutorComponent implements Closeable {
     private Duration maxShutdownWaitTime = Duration.ofSeconds(10);
     @Nullable
     private ExecutorService executor;
+    // also the LOCK object ...
     private final ConcurrentHashMap<TriggerEntity, Future<TriggerKey>> runningTasks = new ConcurrentHashMap<>();
     private final AtomicBoolean stopped = new AtomicBoolean(true);
-    
-    public TaskExecutorComponent(TriggerService triggerService, int maxThreads) {
+
+    public TaskExecutorComponent(String schedulerName, TriggerService triggerService, int maxThreads) {
         super();
+        this.schedulerName = schedulerName;
         this.triggerService = triggerService;
         this.maxThreads.set(maxThreads);
     }
 
     @NonNull
     public List<Future<TriggerKey>> submit(List<TriggerEntity> trigger) {
-        if (trigger == null || trigger.isEmpty()) return Collections.emptyList();
+        if (trigger == null || trigger.isEmpty())
+            return Collections.emptyList();
 
         final List<Future<TriggerKey>> result = new ArrayList<>(trigger.size());
         for (TriggerEntity triggerEntity : trigger) {
@@ -61,10 +70,16 @@ public class TaskExecutorComponent implements Closeable {
         if (trigger == null) {
             return CompletableFuture.completedFuture(null);
         }
-        if (stopped.get()) throw new IllegalStateException("Executor is already stopped");
+        if (stopped.get() || executor == null) {
+            throw new IllegalStateException("Executor of " + schedulerName + " is already stopped");
+        }
 
-        final var result = executor.submit(() -> runTrigger(trigger));
-        runningTasks.put(trigger, result);
+        Future<TriggerKey> result;
+        synchronized (runningTasks) {
+            result = executor.submit(() -> runTrigger(trigger));
+            runningTasks.put(trigger, result);
+        }
+
         return result;
     }
 
@@ -77,22 +92,20 @@ public class TaskExecutorComponent implements Closeable {
         }
     }
 
-    @SuppressWarnings("resource")
-    @PostConstruct
     public void start() {
         if (stopped.compareAndExchange(true, false)) {
-            synchronized (stopped) {
+            synchronized (runningTasks) {
                 runningTasks.clear();
                 executor = Executors.newFixedThreadPool(maxThreads.get());
+                log.info("Started {} with {} threads.", schedulerName, maxThreads.get());
             }
         }
     }
 
     @Override
-    @PreDestroy
     public void close() {
         if (stopped.compareAndExchange(false, true)) {
-            synchronized (stopped) {
+            synchronized (runningTasks) {
                 doShutdown();
             }
         }
@@ -102,8 +115,8 @@ public class TaskExecutorComponent implements Closeable {
         if (executor != null) {
             executor.shutdown();
             if (runningTasks.size() > 0) {
-                log.info("Shutdown executor with {} running tasks, waiting for {}.",
-                        runningTasks.size(), maxShutdownWaitTime);
+                log.info("Shutdown {} with {} running tasks, waiting for {}.", schedulerName, runningTasks.size(),
+                        maxShutdownWaitTime);
 
                 try {
                     executor.awaitTermination(maxShutdownWaitTime.getSeconds(), TimeUnit.SECONDS);
@@ -121,9 +134,16 @@ public class TaskExecutorComponent implements Closeable {
     }
 
     public void shutdownNow() {
-        stopped.set(true);
-        if (executor != null) executor.shutdownNow();
-        executor = null;
+        if (stopped.compareAndExchange(false, true)) {
+            synchronized (runningTasks) {
+                if (executor != null) {
+                    executor.shutdownNow();
+                    log.info("Force stop {} with {} running tasks", schedulerName, runningTasks.size());
+                    runningTasks.clear();
+                    executor = null;
+                }
+            }
+        }
     }
 
     public int getFreeThreads() {
@@ -133,14 +153,18 @@ public class TaskExecutorComponent implements Closeable {
         return Math.max(maxThreads.get() - runningTasks.size(), 0);
     }
 
-    public int getRunningTasks() {
+    public int countRunning() {
         return runningTasks.size();
+    }
+
+    public Collection<Future<TriggerKey>> getRunningTasks() {
+        return runningTasks.values();
     }
 
     public boolean isStopped() {
         return stopped.get() || maxThreads.get() <= 0;
     }
-    
+
     public List<TriggerEntity> getRunningTriggers() {
         return Collections.list(this.runningTasks.keys());
     }
@@ -148,8 +172,9 @@ public class TaskExecutorComponent implements Closeable {
     public void setMaxThreads(int value) {
         this.maxThreads.set(value);
     }
+
     public int getMaxThreads() {
-        return this.maxThreads.get();
+        return isStopped() ? 0 : this.maxThreads.get();
     }
 
     public boolean isRunning(TriggerEntity trigger) {