+++ core/src/main/java/org/sterl/spring/persistent_tasks/trigger/api/TriggerConverter.java
@@ -16,6 +16,7 @@ public class TriggerConverter {
             result.setId(source.getId());
             result.setInstanceId(source.getId());
             result.setRunningOn(source.getRunningOn());
+            result.setLastPing(source.getLastPing());
             return result;
         }
     }