+++ core/src/main/java/org/sterl/spring/persistent_tasks/trigger/component/ReadTriggerComponent.java
@@ -64,7 +64,7 @@ public class ReadTriggerComponent {
         return triggerRepository.findAll(task.name(), page);
     }
 
-    public List<TriggerEntity> findTriggerByCorrelationId(String correlationId) {
-        return triggerRepository.findByCorrelationId(correlationId);
+    public List<TriggerEntity> findTriggerByCorrelationId(String correlationId, Pageable page) {
+        return triggerRepository.findByCorrelationId(correlationId, triggerRepository.applyDefaultSortIfNeeded(page));
     }
 }