+++ test/src/main/java/org/sterl/spring/persistent_tasks/test/PersistentTaskTestService.java
@@ -5,10 +5,14 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.time.Duration;
 import java.time.OffsetDateTime;
 import java.util.ArrayList;
+import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Optional;
+import java.util.Set;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
 
 import org.springframework.stereotype.Service;
 import org.sterl.spring.persistent_tasks.api.TriggerKey;
@@ -17,7 +21,9 @@ import org.sterl.spring.persistent_tasks.scheduler.SchedulerService;
 import org.sterl.spring.persistent_tasks.trigger.TriggerService;
 import org.sterl.spring.persistent_tasks.trigger.model.TriggerEntity;
 
+import lombok.Getter;
 import lombok.RequiredArgsConstructor;
+import lombok.Setter;
 import lombok.SneakyThrows;
 
 @Service
@@ -26,6 +32,9 @@ public class PersistentTaskTestService {
 
     private final List<SchedulerService> schedulers;
     private final TriggerService triggerService;
+    
+    @Getter @Setter
+    private Duration defaultTimeout = Duration.ofSeconds(5);
     
     /**
      * Runs just the next trigger, if it is due to run.
@@ -40,12 +49,12 @@ public class PersistentTaskTestService {
      * Runs all triggers which are due until the given time. One by one, so new triggers are picked up.
      * 
      * @param dueUntil date to also check for trigger in the future
-     * @return the triggeres executed, to directly check if they have been successful
+     * @return the triggers executed, to directly check if they have been successful
      */
     public List<TriggerEntity> runAllDueTrigger(OffsetDateTime dueUntil) {
         var result = new ArrayList<TriggerEntity>();
         List<TriggerEntity> trigger; 
-        while( (trigger = triggerService.lockNextTrigger("test", 1, dueUntil)).size() > 0) {
+        while ( (trigger = triggerService.lockNextTrigger("test", 1, dueUntil)).size() > 0 ) {
             var key = triggerService.run(trigger.getFirst());
             if (key.isPresent()) result.add(key.get());
         }
@@ -53,61 +62,99 @@ public class PersistentTaskTestService {
     }
 
     /**
-     * Triggers the execution of all pending triggers.
+     * Triggers the execution of all pending triggers. We also add one second
+     * to ensure we select all right now created triggers too.
      *
      * @return the reference to the {@link TriggerKey} of the running tasks
      */
     public List<Future<TriggerKey>> scheduleNextTriggers() {
         var result = new ArrayList<Future<TriggerKey>>();
-        if (schedulers.isEmpty()) throw new IllegalStateException("No schedulers found, cannot run any triggers!");
+        assertHasScheduler();
         for (SchedulerService s : schedulers) {
-            result.addAll(s.triggerNextTasks());
+            result.addAll(s.triggerNextTasks(OffsetDateTime.now().plusSeconds(1)));
         }
         return result;
+    }
+
+    public void assertHasScheduler() {
+        assertThat(schedulers).describedAs("No schedulers found, cannot run any triggers!").isNotEmpty();
+    }
+    
+    /**
+     * Triggers the execution of all pending triggers and wait for the result.
+     */
+    @SneakyThrows
+    public Set<TriggerKey> scheduleNextTriggersAndWait() {
+        return scheduleNextTriggersAndWait(defaultTimeout);
     }
 
     /**
      * Triggers the execution of all pending triggers and wait for the result.
      */
     @SneakyThrows
-    public List<TriggerKey> scheduleNextTriggersAndWait(Duration maxWaitTime) {
-        final var result = new ArrayList<TriggerKey>();
+    public Set<TriggerKey> scheduleNextTriggersAndWait(Duration maxWaitTime) {
+        final var result = new LinkedHashSet<TriggerKey>();
         final var timeOut = System.currentTimeMillis() + maxWaitTime.toMillis();
 
-        List<Future<TriggerKey>> triggers;
-        var isSomethingRunning = false;
+        result.addAll(awaitRunningTriggers(maxWaitTime));
+        List<Future<TriggerKey>> newTriggers;
         do {
-            triggers = scheduleNextTriggers();
-            for (Future<TriggerKey> future : triggers) {
-                try {
-                    result.add(future.get());
-                } catch (InterruptedException | ExecutionException e) {
-                    final Throwable cause = e.getCause();
-                    throw cause == null ? e : cause;
-                }
-            }
 
-            isSomethingRunning = hasRunningTriggers();
-            if (isSomethingRunning) {
-                Thread.sleep(Duration.ofMillis(100));
-            }
-            
             if (System.currentTimeMillis() > timeOut) {
                 throw new RuntimeException("Timeout waiting for triggers after " + maxWaitTime);
             }
 
-        } while (!triggers.isEmpty() || isSomethingRunning);
+            newTriggers = scheduleNextTriggers();
+            result.addAll(awaitTriggers(maxWaitTime, newTriggers));
+        } while (newTriggers.size() > 0);
+        return result;
+    }
+    
+    /**
+     * Just waits for the current running triggers
+     * @return return the keys of the currently scheduled triggers
+     */
+    public List<TriggerKey> awaitRunningTriggers() {
+        return awaitRunningTriggers(defaultTimeout);
+    }
 
+    /**
+     * Just waits for the current running triggers
+     * 
+     * @param duration how long to wait
+     * @return return the keys of the currently scheduled triggers
+     */
+    @SneakyThrows
+    public List<TriggerKey> awaitRunningTriggers(Duration duration) {
+        assertHasScheduler();
+        List<Future<TriggerKey>> running = this.schedulers.stream()
+                .flatMap(s -> s.getRunning().stream())
+                .toList();
+        
+        return awaitTriggers(duration, running);
+    }
+
+    public ArrayList<TriggerKey> awaitTriggers(Duration duration, List<Future<TriggerKey>> running) throws Throwable {
+        final var result = new ArrayList<TriggerKey>();
+        final var totalWaitUntil = System.currentTimeMillis() + duration.toMillis();
+        for (Future<TriggerKey> t : running) {
+            try {
+                result.add(t.get(totalWaitUntil - System.currentTimeMillis(), TimeUnit.MILLISECONDS));
+            } catch (InterruptedException | ExecutionException | TimeoutException e) {
+                final Throwable cause = e.getCause();
+                throw cause == null ? e : cause;
+            }
+        }
         return result;
     }
 
-    public boolean hasRunningTriggers() {
-        var running = this.schedulers.stream()
-                .map(s -> s.hasRunningTriggers())
-                .filter(r -> r)
-                .findAny();
+    public int countRunningTriggers() {
+        return schedulers.stream().mapToInt(s -> s.getRunning().size()).sum();
+    }
 
-        return running.isPresent() && running.get() == true;
+    public boolean hasRunningTriggers() {
+        assertHasScheduler();
+        return countRunningTriggers() > 0 ;
     }
 
     public void assertNoMoreTriggers() {