+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
-## v1.6
+## v1.6.1
+-   simpler RetryStrategy - as function
+-   showing last ping
+-   showing execution time or still running triggers
+-   saver way to keep track of running triggers
+
+## v1.6.0 - (2025-03-11)
 
 -   Running triggers can be canceled now
 -   Running triggers can be failed now